third_party/blink/web_tests/external/wpt/editing/run/strikethrough_1-1000-expected.txt

This is a testharness.js-based test.
Found 167 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["strikethrough",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p> <p>bar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:line-through\\">foo</span></p> <p><span style=\\"text-decoration:line-through\\">bar</span></p>" but got "<p><span style=\\"text-decoration-line:line-through\\">foo</span></p> <p><span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<span>[foo</span> <span>bar]</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:line-through\\"><span>foo</span> <span>bar</span></span>" but got "<span style=\\"text-decoration-line:line-through\\"><span>foo</span> <span>bar</span></span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:line-through\\">foo</span></p><p> <span style=\\"text-decoration:line-through\\">bar</span> </p><p><span style=\\"text-decoration:line-through\\">baz</span></p>" but got "<p><span style=\\"text-decoration-line:line-through\\">foo</span></p><p> <span style=\\"text-decoration-line:line-through\\"><span>bar</span> </span></p><p><span style=\\"text-decoration-line:line-through\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><strike>foo</strike></p><p> <strike><span>bar</span></strike> </p><p><strike>baz</strike></p>" but got "<p><strike>foo</strike></p><p> <strike><span>bar</span> </strike></p><p><strike>baz</strike></p>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo<p><br><p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:line-through\\">foo</span></p><p><span style=\\"text-decoration:line-through\\"><br></span></p><p><span style=\\"text-decoration:line-through\\">bar</span></p>" but got "<p><span style=\\"text-decoration-line:line-through\\">foo</span></p><p><span style=\\"text-decoration-line:line-through\\"><br></span></p><p><span style=\\"text-decoration-line:line-through\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<b>foo[]bar</b>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "<b>foo[]bar</b>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<i>foo[]bar</i>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "<i>foo[]bar</i>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar]baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar]baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[bar]baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[bar]baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar<b>baz]qoz</b>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\">bar</span><b><span style=\\"text-decoration:line-through\\">baz</span>qoz</b>quz" but got "foo<span style=\\"text-decoration-line:line-through\\">bar</span><b><span style=\\"text-decoration-line:line-through\\">baz</span>qoz</b>quz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar<i>baz]qoz</i>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\">bar</span><i><span style=\\"text-decoration:line-through\\">baz</span>qoz</i>quz" but got "foo<span style=\\"text-decoration-line:line-through\\">bar</span><i><span style=\\"text-decoration-line:line-through\\">baz</span>qoz</i>quz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p></p><p> </p><p><span style=\\"text-decoration:line-through\\">foo</span></p>" but got "<p></p><p> </p><p><span style=\\"text-decoration-line:line-through\\">foo</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<p><p> <p>foo</p>}" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<p><p> <p>foo</p>}" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "{<p><p> <p>foo</p>}" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "{<p><p> <p>foo</p>}" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td>b<span style=\\"text-decoration:line-through\\">a</span>r</td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td>b<span style=\\"text-decoration-line:line-through\\">a</span>r</td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td><span style=\\"text-decoration:line-through\\">bar</span></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td><span style=\\"text-decoration-line:line-through\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:line-through\\">foo</span></td><td><span style=\\"text-decoration:line-through\\">bar</span></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:line-through\\">foo</span></td><td><span style=\\"text-decoration-line:line-through\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:line-through\\">foo</span></td><td><span style=\\"text-decoration:line-through\\">bar</span></td><td><span style=\\"text-decoration:line-through\\">baz</span></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:line-through\\">foo</span></td><td><span style=\\"text-decoration-line:line-through\\">bar</span></td><td><span style=\\"text-decoration-line:line-through\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:line-through\\">foo</span></td><td><span style=\\"text-decoration:line-through\\">bar</span></td><td><span style=\\"text-decoration:line-through\\">baz</span></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:line-through\\">foo</span></td><td><span style=\\"text-decoration-line:line-through\\">bar</span></td><td><span style=\\"text-decoration-line:line-through\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:line-through\\">foo</span></td><td><span style=\\"text-decoration:line-through\\">bar</span></td><td><span style=\\"text-decoration:line-through\\">baz</span></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:line-through\\">foo</span></td><td><span style=\\"text-decoration-line:line-through\\">bar</span></td><td><span style=\\"text-decoration-line:line-through\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u>[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u>[bar]</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<u>[bar]</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u>[bar]</u>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<u>[bar]</u>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<strike><span style=\\"text-decoration:underline\\">bar</span></strike>baz" but got "foo<u><strike>bar</strike></u>baz"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[bar]baz</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<u>foo<span style=\\"text-decoration:line-through\\">bar</span>baz</u>" but got "<u>foo<span style=\\"text-decoration-line:line-through\\">bar</span>baz</u>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[bar]baz</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[bar]baz</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[bar]baz</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[bar]baz</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<u>foo<span style=\\"text-decoration:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration:line-through\\">ar</span>ba</span>z</u>" but got "<u>foo<span style=\\"text-decoration-line:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration-line:line-through\\">ar</span>ba</span>z</u>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<u>foo<span style=\\"text-decoration:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration:line-through\\">ar</span>ba</span>z</u>" but got "<u>foo<span style=\\"text-decoration-line:line-through\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration-line:line-through\\">ar</span>ba</span>z</u>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<u>foo<span style=\\"text-decoration:line-through\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration:line-through\\">ar</span>ba</span>z</u>" but got "<u>foo<span style=\\"text-decoration-line:line-through\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration-line:line-through\\">ar</span>ba</span>z</u>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<i>ar]ba</i>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<u>foo<span style=\\"text-decoration:line-through\\">b</span><i><span style=\\"text-decoration:line-through\\">ar</span>ba</i>z</u>" but got "<u>foo<span style=\\"text-decoration-line:line-through\\">b</span><i><span style=\\"text-decoration-line:line-through\\">ar</span>ba</i>z</u>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:underline\\">foo<span style=\\"text-decoration:line-through\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:underline\\">foo</span><span style=\\"text-decoration-line:underline line-through\\">bar</span><span style=\\"text-decoration-line:underline\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:underline\\">foo<strike>bar</strike>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:underline\\">foo</span><u><strike>bar</strike>baz</u></p>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[bar]baz</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[bar]baz</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[bar]baz</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[bar]baz</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><s>z</s>" but got "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\">ar<strike>ba</strike></span><s>z</s>" but got "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\">ar<strike>ba</strike></span><strike>z</strike>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><s>z</s>" but got "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<strike>ba</strike></span><s>z</s>" but got "<s>foo</s>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<strike>ba</strike></span><strike>z</strike>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<span style=\\"font-size:3em\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><s>z</s>" but got "<s>foo</s>b<span style=\\"font-size:3em\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></span><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<span style=\\"font-size:3em\\">ar<strike>ba</strike></span><s>z</s>" but got "<s>foo</s>b<span style=\\"font-size:3em\\">ar<strike>ba</strike></span><strike>z</strike>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<i>ar]ba</i>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<i>ar<span style=\\"text-decoration-line:line-through\\">ba</span></i><s>z</s>" but got "<s>foo</s>b<i style=\\"\\">ar<span style=\\"text-decoration-line:line-through\\">ba</span></i><span style=\\"text-decoration-line:line-through\\">z</span>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<i>ar]ba</i>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo</s>b<i>ar<strike>ba</strike></i><s>z</s>" but got "<s>foo</s>b<i style=\\"\\">ar<strike>ba</strike></i><strike>z</strike>"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[bar]baz</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:line-through\\">foo</span>bar<span style=\\"text-decoration:line-through\\">baz</span>" but got "<strike>foo</strike>bar<strike>baz</strike>"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
